Only consider killing Unprotected User's db txns if they are in a transaction and if it was started before $LimitMaxLongestWaitingSeconds seconds
The interval should match the $LimitMaxLongestWaitingSeconds value in the pgterminator source code,
since if the Unprotected User is not in a db txn (xact_start IS NULL)
or if the db txn was started in less than 5 seconds ago,
then they can't be offenders since the Protected Users waiting have been waiting longer than that.

It could happen that the Protected Users were first waiting for something else,
and then continued waiting for some Unprotected User's db txn started more recently than the limit,
but this is hopefully a very uncommon situation, so let's wait and see if it ever happens.